refactor(apps/web): restructure track edit into edit/ subfolder
Made-with: Cursor
